requirements.txt
git+https://github.com/vuvu700/holo.git@dev
pillow
attrs
pynput
screeninfo
numpy